requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aerich==0.7.2
    # via -r requirements.in
aio-pika==9.4.1
    # via -r requirements.in
aioredis==2.0.1
    # via -r requirements.in
aiormq==6.8.0
    # via aio-pika
aiosqlite==0.17.0
    # via tortoise-orm
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
async-timeout==4.0.3
    # via
    #   aioredis
    #   asyncpg
asyncpg==0.29.0
    # via tortoise-orm
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
classy-fastapi==0.6.1
    # via -r requirements.in
click==8.1.7
    # via
    #   aerich
    #   uvicorn
dependency-injector==4.41.0
    # via -r requirements.in
dictdiffer==0.9.0
    # via aerich
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via pydantic
exceptiongroup==1.2.1
    # via
    #   anyio
    #   pytest
fastapi==0.110.3
    # via
    #   -r requirements.in
    #   classy-fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via -r requirements.in
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   yarl
iniconfig==2.0.0
    # via pytest
iso8601==1.1.0
    # via tortoise-orm
multidict==6.0.5
    # via yarl
packaging==24.0
    # via pytest
pamqp==3.3.0
    # via aiormq
passlib==1.7.4
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
pydantic[email]==2.7.1
    # via
    #   -r requirements.in
    #   aerich
    #   classy-fastapi
    #   fastapi
    #   pydantic-settings
    #   tortoise-orm
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
pyjwt==2.8.0
    # via -r requirements.in
pypika-tortoise==0.1.6
    # via tortoise-orm
pytest==8.2.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.23.6
    # via -r requirements.in
python-dotenv==1.0.1
    # via pydantic-settings
python-json-logger==2.0.7
    # via -r requirements.in
pytz==2024.1
    # via tortoise-orm
six==1.16.0
    # via dependency-injector
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
tomli==2.0.1
    # via pytest
tomlkit==0.12.4
    # via aerich
tortoise-orm[asyncpg]==0.20.1
    # via
    #   -r requirements.in
    #   aerich
typing-extensions==4.11.0
    # via
    #   aioredis
    #   aiosqlite
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
uuid6==2024.1.12
    # via -r requirements.in
uvicorn==0.29.0
    # via -r requirements.in
yarl==1.9.4
    # via
    #   aio-pika
    #   aiormq
